fixup! Implement provider-side keymgmt_dup function
