+++ providers/implementations/keymgmt/ecx_kmgmt.c
@@ -694,9 +694,8 @@ void *ecx_load(const void *reference, size_t reference_sz)
 
 static void *ecx_dup(const void *keydata_from)
 {
-    if (ossl_prov_is_running()) {
+    if (ossl_prov_is_running())
         return ossl_ecx_key_dup(keydata_from);
-    }
     return NULL;
 }
