+++ providers/implementations/keymgmt/rsa_kmgmt.c
@@ -648,9 +648,8 @@ static void *rsapss_load(const void *reference, size_t reference_sz)
 
 static void *rsa_dup(const void *keydata_from)
 {
-    if (ossl_prov_is_running()) {
+    if (ossl_prov_is_running())
         return ossl_rsa_dup(keydata_from);
-    }
     return NULL;
 }
